j'ai cast le prix en int

+++ code/common/GameState.cpp
@@ -109,7 +109,7 @@ void GameState::addTower(AbstractTower * tower) {
 void GameState::deleteTower(Position& position, int& quadrant){
     for (AbstractTower *tower : getTowers()){
         if (tower->getPosition() == position){
-            getPlayerStates()[quadrant].earnMoney(tower->getPrice() * PERCENTAGE_RECOVERED_MONEY);
+            getPlayerStates()[quadrant].earnMoney((int) (tower->getPrice() * PERCENTAGE_RECOVERED_MONEY));
             towers.erase(std::remove(getTowers().begin(), getTowers().end(), tower));
         }
     }
